requirements.txt
# Django and REST Framework
Django==5.1.6
djangorestframework==3.15.2

# Data Science & Machine Learning
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.3.2

# NLP (Hugging Face Transformers)
transformers==4.40.0
torch==2.3.1
sentencepiece==0.1.99

# Financial Data APIs
yfinance>=0.2.66
requests==2.31.0

# RSS Feed Parser (optional, for etl/news.py)
feedparser==6.0.11

# Utilities
python-dateutil==2.9.0

# Production Server (for Docker deployment)
gunicorn>=21.2.0
psycopg2-binary>=2.9.9  # For PostgreSQL

